Changed Double to BigDecimal, Added joinColumn and fixing Eli request changes

+++ src/main/java/com/pei/domain/Transaction.java
@@ -1,5 +1,6 @@
 package com.pei.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
@@ -19,12 +20,12 @@ public class Transaction {
     private LocalDateTime date;
 
     @Column(nullable = false)
-    private Double amount;
+    private BigDecimal amount;
 
     public Transaction() {
     }
 
-    public Transaction(Double amount) {
+    public Transaction(BigDecimal amount) {
         this.amount = amount;
     }
 
@@ -41,11 +42,11 @@ public class Transaction {
         this.date = date;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
